+++ ZFXEngine/src/ZFXVector4.cpp
@@ -17,7 +17,7 @@ namespace ZFX
 
 	Vector4 Vector4::Normalize()
 	{
-		float64 f = sqrt(x * x + y * y + z * z);
+		float32 f = sqrt(x * x + y * y + z * z);
 		if (f != 0.0f)
 		{
 			f = 1 / f;
